reduce the number of point lights in shop scene
There were an obscene number of point lights in the shop scene. Or well,
this would not normally be an issue, as the 2d lighting is somewhat
performant to calculate on the GPU, _but_ as we are rendering on WebGPU,
even measly 9 point lights are enough to bring some browser
implementations to their knees.

I thougth 9 point lights wouldn't be that bad, but seems to be
considered 'obscene' for some browser-gpu-driver-whatever combinations.

Incomplete implementations of Web Standards is why we can't have nice
things, I guess.

Reduced the number of point lights and made the remaining ones slightly
brighter. Players likely won't spot a difference.