drive player animations with animationtree
Attempt at decoupling animations from the player movement logic. This
should/could make debugging the animations easier. Hopefully.

The tools of the trade here is an AnimationTree. The tree uses a bunch
of nested state machines to make all possible permutations of different
states smoothly transition to one another. Most of the nested machines
are used just for handling player facing left/right, while the root
level state machine handles moving between the various states.

Some states (e.g. dying/jumping) are not something player is
continuously in, but rather happen once, with player either stuck in
them (dead) or smoothly transferring to other states (jump -> fall).
These are easily modeled using the animation tree, but the initial
transition is a bit tricky.

The jump transition for example needs to be one-off, so having a
continuously truthy condition/expression for it is a bit iffy. Likewise,
death is easiest to detect with the "player died" -signal, instead of
some variable/expression.

Therefore, I ended up making those transitions as "state
teleportations", meaning that there is no transition for these in the
state machine, but rather a signal handler teleports/travels to those
states via code. This is a tradeoff between now having two different
kinds of transitions for the animations and the ease of implementation.

To make matters a bit less confusing, I named these special states with
action_* -prefix, to indicate they may be teleported to.

All-in-all, the animation state machine can now be debugged in-editor,
the animations are driven with a animation tree with two-layer
underlying animation player thingamajig and there is still a script for
defining the various conditions for the animation. Therefore, the
complexity didn't go down, but rather a notch up, which is somewhat
counterintuitive.

However, I argue that the script is now _much_ simpler, and the visual
representation of the AnimationTree is much easier to grasp than the
spaghetti the old 'player_animations.gd' was.

There's likely still a lot to improve, but learning to use the tools
takes time, I guess :)