decouple shooting offset from player animations
To clean up the player scene, it'd be nice to extract the player visuals
to their own scene. However, the animations are currently still strongly
coupled with the physics at some places, making this
difficult/impossible.

Likely the last piece of this puzzle is decoupling the harpoon spawner
position from the player anims. After this, the player_sprite.gd and the
animations it is driving should now be fully self-contained, with just
signals coming in from the parent player.gd.