scene for kill every fish rooms
This is a lie; in truth, all rooms still use this same scene, so
basically this was a really backwards way of recreating the room.tscn
again with different name and built-in objective. However, things have
been shuffled around quite a bit, at least on the code side of things,
so technically we are quite close to separate shop/tutorial room scenes
now.