more dangerous, less annoying pufferfish
The pufferfish now try to select the target positions from near the
player, instead of picking anywhere on the navmesh.

This has some advantages:
1. less chasing for the player, the enemies will come to you. No more
   running around the level, chasing that one pufferfish that keeps on
   camping inside a wall etc.
2. the pufferfish are now actually dangerous if left unchecked. because
   they follow the player around, they are now significantly more likely
   to be in a position where the spikes hit the player

Did a quick test run with these changes and the puffies feel a lot nicer
now; not at all as annoying as before!

Did a quick test run with these changes and the puffies feel a lot nicer
now; not at all as annoying as before!